Hide tooltips via ESC key
Please excuse the hacky way I did this! I'm not super well versed in jQuery plugin development, nor am I very familiar with git, unfortunately. But this seems to work and solved my issue, so I figured I'd share in case it's useful. Added an option to hide tooltips via the ESC key. As I understand it this is a requirement for Success Criterion 1.4.13 of WACG 2.1, I needed this for an project.